Refactor the value of padding and margin
Change the value of padding to 32px to both div one and div three and margin to 12px of .one class So there is space between div one and div two. Change the value of margin bottom of .two class to 48px so there is space between div two and div three. Change the value of left margin of .three class to auto to align div three to the right.